+++ pytos/__init__.py
@@ -1,3 +1,3 @@
 
-__author__ = "Solution Engineering"
-__version__ = "0.0.1"
+__author__ = "PS-Solutions"
+__version__ = "1.0.0"